Fixed the UI not changing color properly
The UI now changes color depending on the color of the owner of the
selected province. This is so that different UI colors can be tried out.
In the future there'll be a setting in the options menu that will allow
users to specify the color of their UI or if they want to set it to
their country's color, or to change depending on what they have
selected.