Added some Japanese provinces + province tool
There's a new tool for creating provinces! The
countrybuilder/CountryBuilder.java class contains the code that is
required for creating a new province with ease. It will put it in the
newly supported res/history/provinces/gen directory, where generated
files are stored. They shouldn't be moved to the regular directory
(res/history/provinces/) until someone has taken a look at them and
determined that they're proper.

It's also possible to reload provinces.

I've added some parts of Japan, currently I'm looking at a system for
recognizing provinces in an easier way than clicking on black parts of
the map and hoping that you're in the right province.