Generated provinces for Japan
Japan is now in the game with a historically accurate population for
1949! Using the new CountryBuilder I built Japan in roughly an hour:
much lesse than what it would have took if I'd done it like I did
previous provinces. Thanks to using EU4 base tax as a guide for
population concentration (not perfect but decent), the process is much,
much faster. Credit to EU4 Extended Timeline mod for providing the names
of provinces and cities as well as their base tax.