Created a new CountryBuilder for colonials
This one doesn't require knowing the basetax beforehand and as such is a
much more useful tool for, for example colonial powers but also to get
more accurate populations. EU4 isn't very accurate in it's
determinations of Base Tax, as both the Select Nation menu and the
in-game Economy view has a Base Tax Sum field. Neither of those is
correct. This way it'll be correct, however.